added nav bar and finished profiles

+++ posts_project/posts/views.py
@@ -5,9 +5,12 @@ from django.http import JsonResponse
 from .forms import PostForm
 from profiles.models import Profile
 from django.http import HttpResponse
+from .utils import action_permission
+from django.contrib.auth.decorators import login_required
+
 
 # Create your views here.
-
+@login_required
 def post_list_and_create(request):
     form = PostForm(request.POST or None)
 
@@ -89,6 +92,7 @@ def post_detail(request, pk):
 
     return render(request, 'posts/detail.html', {'obj': obj, 'form': form})
 
+@action_permission
 def update_post(request, pk):
     obj = Post.objects.get(pk=pk)
 
@@ -103,10 +107,14 @@ def update_post(request, pk):
         'body': obj.body,
     })
 
+@action_permission
 def delete_post(request, pk):
-    obj = Post.objects.get(pk=pk)
-    obj.delete()
-    return JsonResponse({})
+    if request.headers.get('x-requested-with') == 'XMLHttpRequest':
+        obj = Post.objects.get(pk=pk)
+        obj.delete()
+        return JsonResponse({})  
+    
+    return JsonResponse({'msg': 'Denied.'})
 
 def image_upload_view(request):
     if request.method == 'POST':
@@ -114,5 +122,4 @@ def image_upload_view(request):
         new_post_id = request.POST.get('new_post_id')
         post = Post.objects.get(id=new_post_id)
         Photo.objects.create(post=post, image=img)
-
     return HttpResponse()
